fix(types): many more types fixes
refactor: moved ValueOf into types package
