+++ packages/paste-theme/package.json
@@ -28,12 +28,14 @@
     "@emotion/core": "^10.0.14",
     "@emotion/styled": "^10.0.14",
     "@twilio-paste/theme-tokens": "^0.4.0",
+    "@twilio-paste/types": "^0.0.1",
     "react": "^16.8.6",
     "react-dom": "^16.8.6",
     "styled-system": "^4.1.0"
   },
   "devDependencies": {
     "@twilio-paste/theme-tokens": "^0.6.0",
+    "@twilio-paste/types": "^0.0.1",
     "rollup": "^1.16.2",
     "rollup-plugin-babel": "^4.3.3",
     "rollup-plugin-commonjs": "^10.0.1",
