+++ packages/paste-types/src/index.ts
@@ -1,9 +1,4 @@
 export {SendGridThemeShape, ThemeShape} from '@twilio-paste/theme-tokens';
 
-export * from './tokens/backgroundColor';
-export * from './tokens/border';
-export * from './tokens/boxShadow';
-export * from './tokens/font';
-export * from './tokens/sizing';
-export * from './tokens/spacing';
-export * from './tokens/zIndex';
+export * from './utils';
+export * from './style-props';
