+++ packages/paste-types/src/style-props/backgroundColor.ts
@@ -1,0 +1,6 @@
+import {ThemeShape} from '@twilio-paste/theme-tokens';
+import {ResponsiveValue} from 'styled-system';
+
+export interface BackgroundColorProps {
+  backgroundColor?: ResponsiveValue<keyof ThemeShape['backgroundColors']>;
+}
