+++ packages/paste-types/src/style-props/boxShadow.ts
@@ -1,0 +1,6 @@
+import {ThemeShape} from '@twilio-paste/theme-tokens';
+import {ResponsiveValue} from 'styled-system';
+
+export interface BoxShadowProps {
+  boxShadow?: ResponsiveValue<keyof ThemeShape['shadows']>;
+}
