+++ packages/paste-types/src/style-props/index.ts
@@ -1,0 +1,7 @@
+export * from './backgroundColor';
+export * from './border';
+export * from './boxShadow';
+export * from './font';
+export * from './sizing';
+export * from './spacing';
+export * from './zIndex';
