+++ packages/paste-types/src/style-props/zIndex.ts
@@ -1,0 +1,6 @@
+import {ThemeShape} from '@twilio-paste/theme-tokens';
+import {ResponsiveValue} from 'styled-system';
+
+export interface ZIndexProps {
+  zIndex?: ResponsiveValue<keyof ThemeShape['zIndices']>;
+}
