--- packages/paste-types/src/tokens/backgroundColor.ts
@@ -1,5 +1,0 @@
-import {ThemeShape} from '@twilio-paste/theme-tokens';
-
-export interface BackgroundColorProps {
-  backgroundColor?: keyof ThemeShape['backgroundColors'] | Array<keyof ThemeShape['backgroundColors']>;
-}
