--- packages/paste-types/src/tokens/boxShadow.ts
@@ -1,5 +1,0 @@
-import {ThemeShape} from '@twilio-paste/theme-tokens';
-
-export interface BoxShadowProps {
-  boxShadow?: keyof ThemeShape['shadows'] | Array<keyof ThemeShape['shadows']>;
-}
