--- packages/paste-types/src/tokens/zIndex.ts
@@ -1,5 +1,0 @@
-import {ThemeShape} from '@twilio-paste/theme-tokens';
-
-export interface ZIndexProps {
-  zIndex?: keyof ThemeShape['zIndices'] | Array<keyof ThemeShape['zIndices']>;
-}
