--- packages/paste-website/types/utils.ts
@@ -1,1 +1,0 @@
-export type ValueOf<O> = O[keyof O];
